Refactored to move the upsert end point to the AlphabeticalSearchController
